LPS-72818 Split GroupServiceTest class into two different test classes
